fixed float column load issue

+++ src/main/scala/com/samelamin/spark/bigquery/converters/SchemaConverters.scala
@@ -87,7 +87,7 @@ object SchemaConverters {
   def getTypeName(dataType: String):DataType ={
     dataType match {
       case "INTEGER" => LongType
-      case "FLOAT" => FloatType
+      case "FLOAT" => DoubleType
       case "STRING" => StringType
       case "BYTES" => BinaryType
       case "BOOLEAN" => BooleanType
